refactor: 역할에 맞는 패키지로 분류해서 파일 이동
- common: 열거형 객체 이동
- domain
 - entity: Lotto,User,WinningNumbers 등 비즈니스 모델 이동
 - usecase: 각 usecase 이동
- validator: 예외를 처리하는 validator 이동
- ui
 - view: 각종 view 이동
 - controller: 앱의 흐름을 가지는 controller 이동
기능을 분배하면서 많은 클래스가 생성되서 역할에 맞는 패키지로 분류
