+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,14 @@
 package lotto
 
+import lotto.domain.usecase.CalculateLottoRankUseCase
+import lotto.domain.usecase.CalculateLottoReturnUseCase
+import lotto.domain.usecase.CreateUserUseCase
+import lotto.domain.usecase.CreateWinningNumbersUseCase
+import lotto.ui.view.LottoBuyView
+import lotto.ui.controller.LottoController
+import lotto.ui.view.LottoResultView
+import lotto.ui.view.WinningNumberInputView
+
 fun main() {
     val calculateLottoRankUseCase = CalculateLottoRankUseCase()
     val calculateLottoReturnUseCase = CalculateLottoReturnUseCase()
