+++ src/main/kotlin/lotto/domain/entity/User.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain.entity
 
 import camp.nextstep.edu.missionutils.Randoms
