+++ src/main/kotlin/lotto/domain/entity/WinningNumbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain.entity
 
 class WinningNumbers(private val numbers: List<Int>, val bonusNumber: Int) {
     fun getNumbers() = numbers.toList()
