+++ src/main/kotlin/lotto/ui/controller/LottoController.kt
@@ -1,4 +1,15 @@
-package lotto
+package lotto.ui.controller
+
+import lotto.domain.entity.User
+import lotto.domain.entity.Lotto
+import lotto.domain.entity.WinningNumbers
+import lotto.domain.usecase.CalculateLottoRankUseCase
+import lotto.domain.usecase.CalculateLottoReturnUseCase
+import lotto.domain.usecase.CreateUserUseCase
+import lotto.domain.usecase.CreateWinningNumbersUseCase
+import lotto.ui.view.LottoBuyView
+import lotto.ui.view.LottoResultView
+import lotto.ui.view.WinningNumberInputView
 
 class LottoController(
     private val calculateLottoRankUseCase: CalculateLottoRankUseCase,
