+++ src/main/kotlin/lotto/ui/view/WinningNumberInputView.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.ui.view
 
 import camp.nextstep.edu.missionutils.Console
