+++ src/test/kotlin/lotto/UserTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.domain.entity.User
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
